Deployment/requirements.txt
streamlit
pandas
seaborn
numpy
matplotlib
scikit-learn==1.2.2
Pillow
plotly
feature-engine
tensorflow